adapt padding to window operations

+++ src/main/python/systemds/scuro/representations/tfidf.py
@@ -43,8 +43,7 @@ class TfIdf(UnimodalRepresentation):
         vectorizer = TfidfVectorizer(min_df=self.min_df)
 
         X = vectorizer.fit_transform(modality.data)
-        X = [np.array(x).reshape(1, -1) for x in X.toarray()]
-
+        X = [np.array(x).astype(np.float32).reshape(1, -1) for x in X.toarray()]
         if self.output_file is not None:
             save_embeddings(X, self.output_file)
